fix: move default headers
