+++ x-pack/plugins/security_solution/public/common/components/events_viewer/default_alert_headers.ts
@@ -6,7 +6,7 @@
  */
 
 import type { ColumnHeaderOptions } from '../../../../common/types/timeline';
-import { defaultColumnHeaderType } from '../../../timelines/components/timeline/unified_components/default_headers';
+import { defaultColumnHeaderType } from '../../../timelines/components/timeline/body/column_headers/default_headers';
 import { DEFAULT_DATE_COLUMN_MIN_WIDTH } from '../../../timelines/components/timeline/body/constants';
 
 export const defaultAlertsHeaders: ColumnHeaderOptions[] = [
