+++ x-pack/plugins/security_solution/public/common/components/sessions_viewer/default_headers.ts
@@ -8,7 +8,7 @@
 import { tableDefaults } from '@kbn/securitysolution-data-table';
 import type { SubsetDataTableModel } from '@kbn/securitysolution-data-table';
 import type { ColumnHeaderOptions } from '../../../../common/types/timeline';
-import { defaultColumnHeaderType } from '../../../timelines/components/timeline/unified_components/default_headers';
+import { defaultColumnHeaderType } from '../../../timelines/components/timeline/body/column_headers/default_headers';
 import { DEFAULT_DATE_COLUMN_MIN_WIDTH } from '../../../timelines/components/timeline/body/constants';
 import {
   COLUMN_SESSION_START,
