+++ x-pack/plugins/security_solution/public/detections/configurations/security_solution_detections/columns.ts
@@ -15,7 +15,7 @@ import {
 import type { LicenseService } from '../../../../common/license';
 import type { ColumnHeaderOptions } from '../../../../common/types';
 
-import { defaultColumnHeaderType } from '../../../timelines/components/timeline/unified_components/default_headers';
+import { defaultColumnHeaderType } from '../../../timelines/components/timeline/body/column_headers/default_headers';
 import {
   DEFAULT_COLUMN_MIN_WIDTH,
   DEFAULT_DATE_COLUMN_MIN_WIDTH,
