+++ x-pack/plugins/security_solution/public/kubernetes/pages/constants.ts
@@ -6,7 +6,7 @@
  */
 
 import type { ColumnHeaderOptions } from '../../../common/types/timeline';
-import { defaultColumnHeaderType } from '../../timelines/components/timeline/unified_components/default_headers';
+import { defaultColumnHeaderType } from '../../timelines/components/timeline/body/column_headers/default_headers';
 import { DEFAULT_DATE_COLUMN_MIN_WIDTH } from '../../timelines/components/timeline/body/constants';
 import {
   COLUMN_SESSION_START,
