+++ x-pack/plugins/security_solution/public/timelines/components/fields_browser/index.tsx
@@ -19,7 +19,7 @@ import { useKibana } from '../../../common/lib/kibana';
 import { sourcererSelectors } from '../../../common/store';
 import type { State } from '../../../common/store';
 import type { SourcererScopeName } from '../../../sourcerer/store/model';
-import { defaultColumnHeaderType } from '../timeline/unified_components/default_headers';
+import { defaultColumnHeaderType } from '../timeline/body/column_headers/default_headers';
 import { DEFAULT_COLUMN_MIN_WIDTH } from '../timeline/body/constants';
 import { useCreateFieldButton } from './create_field_button';
 import { useFieldTableColumns } from './field_table_columns';
