+++ x-pack/plugins/security_solution/public/timelines/components/modal/actions/new_timeline_button.test.tsx
@@ -12,7 +12,7 @@ import { TimelineId } from '../../../../../common/types';
 import { timelineActions } from '../../../store';
 import { TestProviders } from '../../../../common/mock';
 import { RowRendererValues } from '../../../../../common/api/timeline';
-import { defaultUdtHeaders } from '../../timeline/unified_components/default_headers';
+import { defaultUdtHeaders } from '../../timeline/body/column_headers/default_headers';
 
 jest.mock('../../../../common/components/discover_in_timeline/use_discover_in_timeline_context');
 jest.mock('../../../../common/hooks/use_selector');
