+++ x-pack/plugins/security_solution/public/timelines/components/timeline/tabs/shared/use_timeline_columns.test.ts
@@ -8,7 +8,7 @@
 import { TestProviders } from '../../../../../common/mock';
 import { renderHook } from '@testing-library/react-hooks';
 import { useTimelineColumns } from './use_timeline_columns';
-import { defaultUdtHeaders } from '../../unified_components/default_headers';
+import { defaultUdtHeaders } from '../../body/column_headers/default_headers';
 import type { ColumnHeaderOptions } from '../../../../../../common/types/timeline/columns';
 
 jest.mock('../../../../../common/hooks/use_experimental_features', () => ({
