+++ x-pack/plugins/security_solution/public/timelines/components/timeline/unified_components/index.test.tsx
@@ -32,7 +32,7 @@ import { useIsExperimentalFeatureEnabled } from '../../../../common/hooks/use_ex
 import { TimelineTabs } from '@kbn/securitysolution-data-table';
 import { DataLoadingState } from '@kbn/unified-data-table';
 import { getColumnHeaders } from '../body/column_headers/helpers';
-import { defaultUdtHeaders } from './default_headers';
+import { defaultUdtHeaders } from '../body/column_headers/default_headers';
 import type { ColumnHeaderType } from '../../../../../common/types';
 
 jest.mock('../../../containers', () => ({
@@ -43,10 +43,6 @@ jest.mock('../../../containers/details');
 
 jest.mock('../../fields_browser', () => ({
   useFieldBrowserOptions: jest.fn(),
-}));
-
-jest.mock('../body/events', () => ({
-  Events: () => <></>,
 }));
 
 jest.mock('../../../../sourcerer/containers');
